refactor: app broken up
