+++ src/app.tsx
@@ -1,38 +1,23 @@
 import type { Message } from "@ai-sdk/react";
-import {
-  Bug,
-  Moon,
-  PaperPlaneRight,
-  Stop,
-  Sun,
-  Trash,
-} from "@phosphor-icons/react";
-import { Lightbulb } from "@phosphor-icons/react/dist/ssr";
 import { useAgentChat } from "agents/ai-react";
 import { useAgent } from "agents/react";
-import type React from "react";
-import { useCallback, useEffect, useState, useId } from "react";
-
-import loresmith from "@/assets/loresmith.png";
+import { useEffect, useState, useId } from "react";
 
 // Component imports
-import { Button } from "@/components/button/Button";
-import { Card } from "@/components/card/Card";
-import { HelpButton } from "@/components/help/HelpButton";
-import { MemoizedMarkdown } from "@/components/memoized-markdown";
-
-import { Textarea } from "@/components/textarea/Textarea";
-import { Toggle } from "@/components/toggle/Toggle";
-import { ToolInvocationCard } from "@/components/tool-invocation-card/ToolInvocationCard";
-import { ThinkingSpinner } from "@/components/thinking-spinner";
-import { ResourceSidePanel } from "@/components/resource-side-panel";
-import { BlockingAuthenticationModal } from "./components/BlockingAuthenticationModal";
+import {
+  useAuth,
+  ChatHeader,
+  WelcomeMessage,
+  ChatInput,
+  ChatMessages,
+  useHelpSystem,
+  useSessionManager,
+  ThinkingSpinner,
+  ResourceSidePanel,
+  BlockingAuthenticationModal,
+} from "@/components";
 
 import { useJwtExpiration } from "./hooks/useJwtExpiration";
-import { AuthService } from "./services/auth-service";
-import { JWT_STORAGE_KEY } from "./constants";
-import { API_CONFIG } from "./shared";
-
 import type { campaignTools } from "./tools/campaign";
 import type { generalTools } from "./tools/general";
 import type { fileTools } from "./tools/file";
@@ -52,123 +37,13 @@ const toolsRequiringConfirmation: (
   "deleteFile",
 ];
 
-/**
- * Generate a unique session ID for this browser session
- * This will be used to create a unique Durable Object ID for each session
- */
-function generateSessionId(): string {
-  return `session-${Date.now()}-${Math.random().toString(36).slice(2, 11)}`;
-}
-
-/**
- * Get or create a session ID, persisting it in localStorage
- * This ensures the same session ID is used across browser sessions
- */
-function getSessionId(): string {
-  const existingSessionId = localStorage.getItem("chat-session-id");
-  if (existingSessionId) {
-    return existingSessionId;
-  }
-
-  const newSessionId = generateSessionId();
-  localStorage.setItem("chat-session-id", newSessionId);
-  return newSessionId;
-}
-
 export default function Chat() {
   const chatContainerId = useId();
-  const [theme, setTheme] = useState<"dark" | "light">(() => {
-    // Check localStorage first, default to dark if not found
-    const savedTheme = localStorage.getItem("theme");
-    return (savedTheme as "dark" | "light") || "dark";
-  });
-
   const [showDebug, setShowDebug] = useState(false);
-  const [textareaHeight, setTextareaHeight] = useState("auto");
 
-  // Authentication state management
-  const [showAuthModal, setShowAuthModal] = useState(false);
-  const [username, setUsername] = useState<string>("");
-  const [storedOpenAIKey, setStoredOpenAIKey] = useState<string>("");
-  const [isAuthenticated, setIsAuthenticated] = useState(false);
-  const [showUserMenu, setShowUserMenu] = useState(false);
-
-  // Get stored JWT for user operations
-  const getStoredJwt = useCallback((): string | null => {
-    return localStorage.getItem(JWT_STORAGE_KEY);
-  }, []);
-
-  // Check for stored OpenAI key
-  const checkStoredOpenAIKey = useCallback(async (username: string) => {
-    try {
-      const response = await fetch(
-        `/get-openai-key?username=${encodeURIComponent(username)}`
-      );
-      const result = (await response.json()) as {
-        hasKey?: boolean;
-        apiKey?: string;
-      };
-      if (response.ok && result.hasKey) {
-        setStoredOpenAIKey(result.apiKey || "");
-        setIsAuthenticated(true);
-      } else {
-        // No stored key found, show the auth modal immediately
-        console.log("[App] No stored OpenAI key found for user:", username);
-        console.log("[App] Showing auth modal immediately");
-        setShowAuthModal(true);
-        setIsAuthenticated(false);
-      }
-    } catch (error) {
-      console.error("Error checking stored OpenAI key:", error);
-      // Show modal on error as well
-      console.log("[App] Error checking stored key, showing auth modal");
-      setShowAuthModal(true);
-      setIsAuthenticated(false);
-    }
-  }, []);
-
-  // Log authentication state changes for debugging
-  useEffect(() => {
-    console.log("[App] Authentication state changed:", isAuthenticated);
-  }, [isAuthenticated]);
-
-  // Check authentication status on mount
-  useEffect(() => {
-    console.log("[App] useEffect running - checking authentication status");
-    const jwt = getStoredJwt();
-    if (jwt) {
-      try {
-        const payload = JSON.parse(atob(jwt.split(".")[1]));
-        if (payload.username) {
-          setUsername(payload.username);
-          // Check if JWT is expired
-          if (AuthService.isJwtExpired(jwt)) {
-            // JWT expired, show auth modal
-            console.log("[App] JWT expired, showing auth modal");
-            setShowAuthModal(true);
-            setIsAuthenticated(false);
-          } else {
-            // JWT valid, check if we have stored OpenAI key
-            console.log("[App] JWT valid, checking stored OpenAI key");
-            checkStoredOpenAIKey(payload.username);
-          }
-        }
-      } catch (error) {
-        console.error("Error parsing JWT:", error);
-        console.log("[App] Error parsing JWT, showing auth modal");
-        setShowAuthModal(true);
-        setIsAuthenticated(false);
-      }
-    } else {
-      // No JWT, show auth modal
-      console.log("[App] No JWT, showing auth modal");
-      setShowAuthModal(true);
-      setIsAuthenticated(false);
-    }
-  }, [checkStoredOpenAIKey, getStoredJwt]);
-
-  // Get session ID for this browser session
-  const sessionId = getSessionId();
+  // Use extracted hooks
+  const auth = useAuth();
+  const { sessionId, handleClearHistory } = useSessionManager();
 
   // Handle JWT expiration globally
   useJwtExpiration({
@@ -176,132 +51,6 @@ export default function Chat() {
       // JWT expired - no annoying toasts needed
     },
   });
-
-  // Handle authentication submission
-  const handleAuthenticationSubmit = async (
-    username: string,
-    adminKey: string,
-    openaiApiKey: string
-  ) => {
-    try {
-      const response = await fetch(
-        API_CONFIG.buildUrl(API_CONFIG.ENDPOINTS.AUTH.AUTHENTICATE),
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({
-            username,
-            adminSecret: adminKey?.trim() || undefined, // Make admin key optional
-            openaiApiKey,
-          }),
-        }
-      );
-
-      const result = (await response.json()) as {
-        success?: boolean;
-        token?: string;
-        error?: string;
-      };
-
-      if (response.ok && result.token) {
-        // Store JWT token
-        AuthService.storeJwt(result.token);
-
-        // Update stored OpenAI key
-        setStoredOpenAIKey(openaiApiKey);
-
-        // Set authentication state
-        setIsAuthenticated(true);
-
-        // Close modal
-        setShowAuthModal(false);
-      } else {
-        throw new Error(result.error || "Authentication failed");
-      }
-    } catch (error) {
-      console.error("Error during authentication:", error);
-      throw error;
-    }
-  };
-
-  useEffect(() => {
-    // Apply theme class on mount and when theme changes
-    if (theme === "dark") {
-      document.documentElement.classList.add("dark");
-      document.documentElement.classList.remove("light");
-    } else {
-      document.documentElement.classList.remove("dark");
-      document.documentElement.classList.add("light");
-    }
-
-    // Save theme preference to localStorage
-    localStorage.setItem("theme", theme);
-  }, [theme]);
-
-  const toggleTheme = () => {
-    const newTheme = theme === "dark" ? "light" : "dark";
-    setTheme(newTheme);
-  };
-
-  const handleLogout = async () => {
-    try {
-      // Call the logout endpoint
-      const response = await fetch(
-        API_CONFIG.buildUrl(API_CONFIG.ENDPOINTS.AUTH.LOGOUT),
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-        }
-      );
-
-      if (response.ok) {
-        // Clear local JWT storage
-        AuthService.clearJwt();
-
-        // Reset authentication state
-        setIsAuthenticated(false);
-        setUsername("");
-        setShowUserMenu(false);
-
-        // Show success message
-        console.log("Logged out successfully");
-
-        // Optionally show auth modal again
-        setShowAuthModal(true);
-      } else {
-        throw new Error("Logout failed");
-      }
-    } catch (error) {
-      console.error("Logout error:", error);
-      console.error("Logout failed. Please try again.");
-
-      // Force clear local state even if server call failed
-      AuthService.clearJwt();
-      setIsAuthenticated(false);
-      setUsername("");
-      setShowUserMenu(false);
-      setShowAuthModal(true);
-    }
-  };
-
-  // Close user menu when clicking outside
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (
-        showUserMenu &&
-        !(event.target as Element).closest(".user-menu-container")
-      ) {
-        setShowUserMenu(false);
-      }
-    };
-
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => document.removeEventListener("mousedown", handleClickOutside);
-  }, [showUserMenu]);
 
   const agent = useAgent({
     agent: "chat",
@@ -330,7 +79,7 @@ export default function Chat() {
         console.log(
           "[App] Authentication required detected, showing auth modal"
         );
-        setShowAuthModal(true);
+        auth.setShowAuthModal(true);
       } else {
         console.log("[App] No authentication required in response");
       }
@@ -363,7 +112,7 @@ export default function Chat() {
         console.log(
           "[App] Authentication required error detected, showing auth modal"
         );
-        setShowAuthModal(true);
+        auth.setShowAuthModal(true);
       } else {
         console.log("[App] Different error type, not showing auth modal");
       }
@@ -395,12 +144,12 @@ export default function Chat() {
 
   // Debug modal state changes
   useEffect(() => {
-    console.log("[App] showAuthModal changed to:", showAuthModal);
-  }, [showAuthModal]);
+    console.log("[App] showAuthModal changed to:", auth.showAuthModal);
+  }, [auth.showAuthModal]);
 
   // Function to handle suggested prompts
   const handleSuggestionSubmit = (suggestion: string) => {
-    const jwt = getStoredJwt();
+    const jwt = auth.getStoredJwt?.();
     console.log("[App] handleSuggestionSubmit sending JWT:", jwt);
 
     // Always send the message to the agent - let the agent handle auth requirements
@@ -423,101 +172,17 @@ export default function Chat() {
   };
 
   // Enhanced clear history function that creates a new session
-  const handleClearHistory = () => {
+  const handleClearHistoryWithReload = () => {
     clearHistory();
-    // Optionally create a new session ID when clearing history
-    // This creates a completely fresh chat session
-    const newSessionId = generateSessionId();
-    localStorage.setItem("chat-session-id", newSessionId);
-    // Reload the page to reinitialize with the new session ID
-    window.location.reload();
+    handleClearHistory();
   };
 
-  // Handle help button actions
-  const handleHelpAction = (action: string) => {
-    const jwt = getStoredJwt();
-    console.log("[App] handleHelpAction:", action);
-
-    let response = "";
-    switch (action) {
-      case "upload_resource":
-        response =
-          "## 📁 Uploading Resources\n\n" +
-          "To upload resources to your inspiration library:\n\n" +
-          "1. **Look for the 'Add to library' button** in the interface\n" +
-          "2. **Click the button** to open the upload modal\n" +
-          "3. **Drag and drop files** directly onto the upload area for quick upload\n" +
-          "4. **Select files** from your computer if you prefer\n\n" +
-          "**Supported file types:** PDF files, images, and other documents\n\n" +
-          "Once uploaded, your resources will be available in your inspiration library for campaign planning!";
-        break;
-      case "create_campaign":
-        response =
-          "## 🎲 Creating a Campaign\n\n" +
-          "To create a new campaign:\n\n" +
-          "1. **Look for the 'Create Campaign' button** in the interface\n" +
-          "2. **Click the button** to open the campaign creation form\n" +
-          "3. **Enter campaign details** including:\n" +
-          "- Campaign name\n" +
-          "- Description\n" +
-          "- Setting details\n" +
-          "4. **Save your campaign** to start organizing your resources\n\n" +
-          "**Benefits:** Campaigns help you organize your resources, plan sessions, and track your story development!";
-        break;
-      case "start_chat":
-        response =
-          "## 💬 Starting a Chat\n\n" +
-          "You can start chatting with me right here! Just type your questions about:\n\n" +
-          "**Campaign Ideas:**\n" +
-          "- World building concepts\n" +
-          "- Plot development\n" +
-          "- Character creation\n\n" +
-          "**GM Topics:**\n" +
-          "- Session planning\n" +
-          "- Encounter design\n" +
-          "- Story pacing\n\n" +
-          "**Tips:**\n" +
-          "- Be specific with your questions\n" +
-          "- Share your campaign context\n" +
-          "- Ask for examples or suggestions\n\n" +
-          "I'm here to help you develop your campaign ideas and provide guidance!";
-        break;
-      default:
-        response =
-          "## 🎯 Getting Started\n\n" +
-          "I can help you with various tasks:\n\n" +
-          "**📁 Upload Resources:**\n" +
-          "- Look for the 'Add to library' button\n" +
-          "- Upload PDFs, images, and documents\n\n" +
-          "**🎲 Create Campaigns:**\n" +
-          "- Use the 'Create Campaign' button\n" +
-          "- Organize your story elements\n\n" +
-          "**💬 Start Chatting:**\n" +
-          "- Just type your questions here\n" +
-          "- Ask about campaign ideas, world building, or GM topics\n\n" +
-          "**💡 Pro Tip:** Be specific with your questions to get the most helpful responses!";
-    }
-
-    // Add the help response as an assistant message
-    append({
-      role: "assistant",
-      content: response,
-      data: jwt ? { jwt } : undefined,
-    });
-    setInput("");
-  };
-
-  // Scroll to bottom when messages change
-  useEffect(() => {
-    // Only scroll if there are messages and we're not in the initial load
-    if (agentMessages.length > 0 && !isLoading) {
-      // Add a small delay to ensure the messages are rendered
-      const timer = setTimeout(() => {
-        // messagesEndRef.current?.scrollIntoView({ behavior: "smooth" }); // This line is removed
-      }, 100);
-      return () => clearTimeout(timer);
-    }
-  }, [agentMessages.length, isLoading]);
+  // Use help system hook
+  const { handleHelpAction } = useHelpSystem({
+    append,
+    setInput,
+    getStoredJwt: auth.getStoredJwt || (() => null),
+  });
 
   const pendingToolCallConfirmation = agentMessages.some((m: Message) =>
     m.parts?.some(
@@ -533,16 +198,12 @@ export default function Chat() {
     )
   );
 
-  const formatTime = (date: Date) => {
-    return date.toLocaleTimeString([], { hour: "2-digit", minute: "2-digit" });
-  };
-
   // Enhanced form submission handler that includes JWT
   const handleFormSubmit = (e: React.FormEvent) => {
     e.preventDefault();
     if (!agentInput.trim()) return;
 
-    const jwt = getStoredJwt();
+    const jwt = auth.getStoredJwt?.();
     console.log("[App] handleFormSubmit sending JWT:", jwt);
 
     // Always send the message to the agent - let the agent handle auth requirements
@@ -553,7 +214,6 @@ export default function Chat() {
       data: jwt ? { jwt } : undefined,
     });
     setInput("");
-    setTextareaHeight("auto"); // Reset height after submission
     // Scroll to bottom after user sends a message
     setTimeout(() => {
       const chatContainer = document.getElementById(chatContainerId);
@@ -572,7 +232,7 @@ export default function Chat() {
       e.preventDefault();
       if (!agentInput.trim()) return;
 
-      const jwt = getStoredJwt();
+      const jwt = auth.getStoredJwt?.();
       console.log("[App] handleKeyDown sending JWT:", jwt);
 
       // Always send the message to the agent - let the agent handle auth requirements
@@ -583,7 +243,6 @@ export default function Chat() {
         data: jwt ? { jwt } : undefined,
       });
       setInput("");
-      setTextareaHeight("auto"); // Reset height on Enter submission
       // Scroll to bottom after user sends a message
       setTimeout(() => {
         const chatContainer = document.getElementById(chatContainerId);
@@ -603,64 +262,21 @@ export default function Chat() {
         <div className="h-[calc(100vh-2rem)] w-full mx-auto max-w-[1400px] flex shadow-xl rounded-md overflow-hidden relative border border-neutral-300 dark:border-neutral-800">
           {/* Resource Side Panel */}
           <ResourceSidePanel
-            isAuthenticated={isAuthenticated}
-            username={username}
-            onLogout={handleLogout}
-            showUserMenu={showUserMenu}
-            setShowUserMenu={setShowUserMenu}
+            isAuthenticated={auth.isAuthenticated}
+            username={auth.username}
+            onLogout={auth.handleLogout}
+            showUserMenu={auth.showUserMenu}
+            setShowUserMenu={auth.setShowUserMenu}
           />
 
           {/* Main Chat Area */}
           <div className="flex-1 flex flex-col">
-            <div className="px-4 py-3 border-b border-neutral-300 dark:border-neutral-800 flex items-center gap-3 sticky top-0 z-10">
-              <div
-                className="flex items-center justify-center rounded-lg"
-                style={{ width: 28, height: 28 }}
-              >
-                <img
-                  src={loresmith}
-                  alt="LoreSmith logo"
-                  width={28}
-                  height={28}
-                  className="object-contain"
-                />
-              </div>
-
-              <div className="flex-1">
-                <h2 className="font-semibold text-base">LoreSmith</h2>
-              </div>
-
-              <div className="flex items-center gap-2 mr-2">
-                <Bug size={16} />
-                <Toggle
-                  toggled={showDebug}
-                  aria-label="Toggle debug mode"
-                  onClick={() => setShowDebug((prev) => !prev)}
-                />
-              </div>
-
-              <HelpButton onActionClick={handleHelpAction} />
-
-              <Button
-                variant="ghost"
-                size="md"
-                shape="square"
-                className="rounded-full h-9 w-9"
-                onClick={toggleTheme}
-              >
-                {theme === "dark" ? <Sun size={20} /> : <Moon size={20} />}
-              </Button>
-
-              <Button
-                variant="ghost"
-                size="md"
-                shape="square"
-                className="rounded-full h-9 w-9"
-                onClick={handleClearHistory}
-              >
-                <Trash size={20} />
-              </Button>
-            </div>
+            <ChatHeader
+              showDebug={showDebug}
+              setShowDebug={setShowDebug}
+              onHelpAction={handleHelpAction}
+              onClearHistory={handleClearHistoryWithReload}
+            />
 
             {/* Main Content Area */}
             <div
@@ -668,244 +284,47 @@ export default function Chat() {
               className="flex-1 overflow-y-auto px-6 py-4 space-y-4 pb-32 max-h-[calc(100vh-10rem)]"
             >
               {agentMessages.length === 0 && (
-                <div className="h-full flex items-center justify-center">
-                  <Card className="p-6 max-w-2xl mx-auto bg-neutral-100 dark:bg-neutral-900">
-                    <div className="text-left space-y-4">
-                      <div className="bg-[#F48120]/10 text-[#F48120] rounded-full p-3 inline-flex">
-                        <img
-                          src={loresmith}
-                          alt="LoreSmith logo"
-                          width={48}
-                          height={48}
-                        />
-                      </div>
-                      <h3 className="font-semibold text-lg">
-                        👋 Welcome to LoreSmith Campaign Planner!
-                      </h3>
-                      <p className="text-muted-foreground text-sm">
-                        Speak your query, and I shall summon the most fitting
-                        agent. Try:
-                      </p>
-                      <div className="flex justify-center gap-2 pt-2">
-                        <Button
-                          size="sm"
-                          variant="secondary"
-                          className="bg-neutral-100 dark:bg-neutral-600"
-                          onClick={() => handleSuggestionSubmit("Get started")}
-                        >
-                          <Lightbulb size={12} />
-                          Get started
-                        </Button>
-                        <Button
-                          size="sm"
-                          variant="secondary"
-                          className="bg-neutral-100 dark:bg-neutral-600"
-                          onClick={() => handleSuggestionSubmit("Show agents")}
-                        >
-                          Show agents
-                        </Button>
-                      </div>
-                    </div>
-                  </Card>
-                </div>
+                <WelcomeMessage onSuggestionSubmit={handleSuggestionSubmit} />
               )}
 
-              {agentMessages
-                .filter((m: Message) => {
-                  // Hide "Get started" messages from display
-                  if (m.role === "user" && m.content === "Get started") {
-                    return false;
-                  }
-                  return true;
-                })
-                .map((m: Message, _index) => {
-                  const isUser = m.role === "user";
-
-                  return (
-                    <div key={m.id}>
-                      {showDebug && (
-                        <pre className="text-xs text-muted-foreground overflow-scroll">
-                          {JSON.stringify(
-                            {
-                              ...m,
-                              parts: m.parts?.filter(
-                                (part) => part.type !== "tool-invocation"
-                              ),
-                            },
-                            null,
-                            2
-                          )}
-                        </pre>
-                      )}
-                      <div
-                        className={`flex ${isUser ? "justify-end" : "justify-start"}`}
-                      >
-                        <div
-                          className={`${
-                            isUser
-                              ? "flex flex-row-reverse gap-2 max-w-[85%]"
-                              : "w-full"
-                          }`}
-                        >
-                          <div className={isUser ? "flex-1" : "w-full"}>
-                            <div>
-                              {m.parts?.map((part, i) => {
-                                if (part.type === "text") {
-                                  return (
-                                    // biome-ignore lint/suspicious/noArrayIndexKey: immutable index
-                                    <div key={i}>
-                                      <Card
-                                        className={`p-3 rounded-md bg-neutral-100 dark:bg-neutral-900 ${
-                                          isUser
-                                            ? "rounded-br-none"
-                                            : "rounded-bl-none border-assistant-border"
-                                        } ${
-                                          part.text.startsWith(
-                                            "scheduled message"
-                                          )
-                                            ? "border-accent/50"
-                                            : ""
-                                        } relative`}
-                                      >
-                                        {part.text.startsWith(
-                                          "scheduled message"
-                                        ) && (
-                                          <span className="absolute -top-3 -left-2 text-base">
-                                            🕒
-                                          </span>
-                                        )}
-                                        <MemoizedMarkdown
-                                          content={part.text.replace(
-                                            /^scheduled message: /,
-                                            ""
-                                          )}
-                                        />
-                                      </Card>
-                                      <p
-                                        className={`text-xs text-muted-foreground mt-1 ${
-                                          isUser ? "text-right" : "text-left"
-                                        }`}
-                                      >
-                                        {formatTime(
-                                          new Date(
-                                            m.createdAt as unknown as string
-                                          )
-                                        )}
-                                      </p>
-                                    </div>
-                                  );
-                                }
-
-                                if (part.type === "tool-invocation") {
-                                  const toolInvocation = part.toolInvocation;
-                                  const toolCallId = toolInvocation.toolCallId;
-                                  const needsConfirmation =
-                                    toolsRequiringConfirmation.includes(
-                                      toolInvocation.toolName as
-                                        | keyof typeof generalTools
-                                        | keyof typeof campaignTools
-                                        | keyof typeof fileTools
-                                    );
-
-                                  // Skip rendering the card when debug is off
-                                  if (!showDebug) return null;
-
-                                  return (
-                                    <ToolInvocationCard
-                                      // biome-ignore lint/suspicious/noArrayIndexKey: using index is safe here as the array is static
-                                      key={`${toolCallId}-${i}`}
-                                      toolInvocation={toolInvocation}
-                                      toolCallId={toolCallId}
-                                      needsConfirmation={needsConfirmation}
-                                      addToolResult={addToolResult}
-                                      showDebug={showDebug}
-                                    />
-                                  );
-                                }
-                                return null;
-                              })}
-                            </div>
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  );
-                })}
+              <ChatMessages
+                messages={agentMessages}
+                showDebug={showDebug}
+                toolsRequiringConfirmation={toolsRequiringConfirmation}
+                addToolResult={addToolResult}
+              />
 
               {/* Thinking Spinner - shown when agent is processing */}
               {isLoading && (
                 <div className="flex justify-start">
                   <div className="w-full">
-                    <Card className="p-3 rounded-md bg-neutral-100 dark:bg-neutral-900 rounded-bl-none border-assistant-border">
+                    <div className="p-3 rounded-md bg-neutral-100 dark:bg-neutral-900 rounded-bl-none border-assistant-border">
                       <ThinkingSpinner />
-                    </Card>
+                    </div>
                   </div>
                 </div>
               )}
             </div>
 
             {/* Input Area */}
-            <form
+            <ChatInput
+              value={agentInput}
+              onChange={handleAgentInputChange}
               onSubmit={handleFormSubmit}
-              className="p-3 bg-neutral-50 border-t border-neutral-300 dark:border-neutral-800 dark:bg-neutral-900"
-            >
-              <div className="flex items-center gap-2">
-                <div className="flex-1 relative">
-                  <Textarea
-                    disabled={pendingToolCallConfirmation}
-                    placeholder={
-                      pendingToolCallConfirmation
-                        ? "Please respond to the tool confirmation above..."
-                        : "What knowledge do you seek today?"
-                    }
-                    className="flex w-full border border-neutral-200 dark:border-neutral-700 px-3 py-2 text-base ring-offset-background placeholder:text-neutral-500 dark:placeholder:text-neutral-400 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-neutral-300 dark:focus-visible:ring-neutral-700 focus-visible:ring-offset-2 dark:focus-visible:ring-offset-neutral-900 disabled:cursor-not-allowed disabled:opacity-50 md:text-sm min-h-[24px] max-h-[calc(75dvh)] overflow-hidden resize-none rounded-2xl !text-base pb-10 dark:bg-neutral-900"
-                    value={agentInput}
-                    onChange={(e) => {
-                      handleAgentInputChange(e);
-                      // Auto-resize the textarea
-                      e.target.style.height = "auto";
-                      e.target.style.height = `${e.target.scrollHeight}px`;
-                      setTextareaHeight(`${e.target.scrollHeight}px`);
-                    }}
-                    onKeyDown={handleKeyDown}
-                    rows={2}
-                    style={{ height: textareaHeight }}
-                  />
-                  <div className="absolute bottom-0 right-0 p-2 w-fit flex flex-row justify-end">
-                    {isLoading ? (
-                      <button
-                        type="button"
-                        onClick={stop}
-                        className="inline-flex items-center cursor-pointer justify-center gap-2 whitespace-nowrap text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg]:size-4 [&_svg]:shrink-0 bg-primary text-primary-foreground hover:bg-primary/90 rounded-full p-1.5 h-fit border border-neutral-200 dark:border-neutral-800"
-                        aria-label="Stop generation"
-                      >
-                        <Stop size={16} />
-                      </button>
-                    ) : (
-                      <button
-                        type="submit"
-                        className="inline-flex items-center cursor-pointer justify-center gap-2 whitespace-nowrap text-sm font-medium ring-offset-background transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg]:size-4 [&_svg]:shrink-0 bg-primary text-primary-foreground hover:bg-primary/90 rounded-full p-1.5 h-fit border border-neutral-200 dark:border-neutral-800"
-                        disabled={
-                          pendingToolCallConfirmation || !agentInput.trim()
-                        }
-                        aria-label="Send message"
-                      >
-                        <PaperPlaneRight size={16} />
-                      </button>
-                    )}
-                  </div>
-                </div>
-              </div>
-            </form>
+              onKeyDown={handleKeyDown}
+              isLoading={isLoading}
+              onStop={stop}
+              pendingToolCallConfirmation={pendingToolCallConfirmation}
+            />
           </div>
         </div>
       </div>
 
       <BlockingAuthenticationModal
-        isOpen={showAuthModal}
-        username={username}
-        storedOpenAIKey={storedOpenAIKey}
-        onSubmit={handleAuthenticationSubmit}
+        isOpen={auth.showAuthModal}
+        username={auth.username}
+        storedOpenAIKey={auth.storedOpenAIKey}
+        onSubmit={auth.handleAuthenticationSubmit}
       />
     </>
   );
