+++ src/components/index.ts
@@ -1,0 +1,29 @@
+// Auth components
+export { useAuth } from "./auth/AuthProvider";
+
+// Theme components
+export { useTheme } from "./theme/ThemeProvider";
+
+// Chat components
+export { ChatHeader } from "./chat/ChatHeader";
+export { WelcomeMessage } from "./chat/WelcomeMessage";
+export { ChatInput } from "./chat/ChatInput";
+export { ChatMessages } from "./chat/ChatMessages";
+
+// Help components
+export { useHelpSystem } from "./help/HelpSystem";
+
+// Session components
+export { useSessionManager } from "./session/SessionManager";
+
+// Existing components
+export { Button } from "./button/Button";
+export { Card } from "./card/Card";
+export { HelpButton } from "./help/HelpButton";
+export { MemoizedMarkdown } from "./memoized-markdown";
+export { Textarea } from "./textarea/Textarea";
+export { Toggle } from "./toggle/Toggle";
+export { ToolInvocationCard } from "./tool-invocation-card/ToolInvocationCard";
+export { ThinkingSpinner } from "./thinking-spinner";
+export { ResourceSidePanel } from "./resource-side-panel";
+export { BlockingAuthenticationModal } from "./BlockingAuthenticationModal";
