+++ src/components/session/SessionManager.tsx
@@ -1,0 +1,42 @@
+import { useCallback } from "react";
+
+/**
+ * Generate a unique session ID for this browser session
+ * This will be used to create a unique Durable Object ID for each session
+ */
+function generateSessionId(): string {
+  return `session-${Date.now()}-${Math.random().toString(36).slice(2, 11)}`;
+}
+
+/**
+ * Get or create a session ID, persisting it in localStorage
+ * This ensures the same session ID is used across browser sessions
+ */
+function getSessionId(): string {
+  const existingSessionId = localStorage.getItem("chat-session-id");
+  if (existingSessionId) {
+    return existingSessionId;
+  }
+
+  const newSessionId = generateSessionId();
+  localStorage.setItem("chat-session-id", newSessionId);
+  return newSessionId;
+}
+
+export function useSessionManager() {
+  const sessionId = getSessionId();
+
+  const handleClearHistory = useCallback(() => {
+    // Optionally create a new session ID when clearing history
+    // This creates a completely fresh chat session
+    const newSessionId = generateSessionId();
+    localStorage.setItem("chat-session-id", newSessionId);
+    // Reload the page to reinitialize with the new session ID
+    window.location.reload();
+  }, []);
+
+  return {
+    sessionId,
+    handleClearHistory,
+  };
+}
