+++ src/components/theme/ThemeProvider.tsx
@@ -1,0 +1,40 @@
+import { useEffect, useState } from "react";
+
+export type Theme = "dark" | "light";
+
+export interface ThemeContextType {
+  theme: Theme;
+  toggleTheme: () => void;
+}
+
+export function useTheme(): ThemeContextType {
+  const [theme, setTheme] = useState<Theme>(() => {
+    // Check localStorage first, default to dark if not found
+    const savedTheme = localStorage.getItem("theme");
+    return (savedTheme as Theme) || "dark";
+  });
+
+  useEffect(() => {
+    // Apply theme class on mount and when theme changes
+    if (theme === "dark") {
+      document.documentElement.classList.add("dark");
+      document.documentElement.classList.remove("light");
+    } else {
+      document.documentElement.classList.remove("dark");
+      document.documentElement.classList.add("light");
+    }
+
+    // Save theme preference to localStorage
+    localStorage.setItem("theme", theme);
+  }, [theme]);
+
+  const toggleTheme = () => {
+    const newTheme = theme === "dark" ? "light" : "dark";
+    setTheme(newTheme);
+  };
+
+  return {
+    theme,
+    toggleTheme,
+  };
+}
